FAT-5006 (#1012)
* changed patron blocks

* fix bug

* increased waiter

* fixed bug

* increased waiter

* fixed bug

* fixed C9217

* change waiter
